# backend/swagger.yaml
openapi: 3.0.0
info:
  title: TourGuide AI RAG API
  version: 2.0.0
  description: |
    Multimodal AI system with:

    -  JWT Authentication
    -  Multi-chat system (chatId आधारित)
    -  RAG-based responses
    -  Image understanding + contextual chat
    -  MongoDB persistent storage

servers:
  - url: http://localhost:3000/api

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

tags:
  - name: Auth
  - name: Chat
  - name: Image

paths:
  
  # ---------------- ADMIN AUTH ----------------
  /admin/auth/register:
    post:
      summary: Register a new Admin (Requires Secret Key)
      tags: [Admin Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, secret]
              properties:
                username:
                  type: string
                password:
                  type: string
                secret:
                  type: string
                  description: The system-wide ADMIN_SECRET key
      responses:
        201:
          description: Admin created
        401:
          description: Invalid secret key

  /admin/auth/login:
    post:
      summary: Login as Admin to get JWT token
      tags: [Admin Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string

  # ---------------- ADMIN MANAGEMENT ----------------

  /admin/files:
    get:
      summary: List all uploaded files
      tags: [Admin Management]
      security:
        - bearerAuth: [] # Assuming you use Bearer tokens for verifyAdmin
      responses:
        '200':
          description: A successful response returning an array of filenames.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Files retrieved successfully"
                  files:
                    type: array
                    items:
                      type: string
                    example: ["1715638200000-document.pdf", "1715638299999-data.txt"]
        '401':
          description: Unauthorized - Admin access required.
        '500':
          description: Server error or directory not found.


  /admin/upload-and-ingest:
    post:
      summary: Upload document and trigger RAG ingestion
      tags: [Admin Management]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Document to be ingested into the vector database
      responses:
        200:
          description: File uploaded and ingestion complete
        401:
          description: Unauthorized (Invalid or missing token)

  # ---------------- AUTH USER ----------------
  /auth/register:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  example: dilshan
                email:
                  type: string
                  format: email
                  example: dilshan@gmail.com
                password:
                  type: string
                  example: 123456
      responses:
        200:
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User registered successfully

  /auth/login:
    post:
      summary: Login user and get JWT token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string

  # ---------------- CHAT ----------------

  /chat/create:
    post:
      summary: Create a new chat
      tags: [Chat]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Chat created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  chatId:
                    type: string

  /chat:
    get:
      summary: Get all chats for logged-in user
      tags: [Chat]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of chats
          content:
            application/json:
              schema:
                type: object
                properties:
                  chats:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        title:
                          type: string
                        createdAt:
                          type: string

  /chat/{chatId}:
    get:
      summary: Get chat history
      tags: [Chat]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Chat history
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                  imageContext:
                    type: string
                    nullable: true

    post:
      summary: Send message in a chat
      tags: [Chat]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  example: Tell me about Sigiriya
      responses:
        200:
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                  history:
                    type: array
                  imageContext:
                    type: string
                    nullable: true

  /chat/{chatId}/image-context:
    post:
      summary: Set image context for a chat
      tags: [Chat]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [imageContext]
              properties:
                imageContext:
                  type: string
                  example: Sigiriya Rock Fortress
      responses:
        200:
          description: Context stored

  # ---------------- IMAGE ----------------

  /image-search:
    post:
      summary: Upload image and get AI + RAG response
      tags: [Image]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        200:
          description: Image processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  topic:
                    type: string
                  identifiedSubject:
                    type: string
                  data:
                    type: object